Add click listener and some refactoring
Now it's possible to click on te cluster and the item cluster to perform some actions.
A third party library Anko is added in order to handle background task easily. A little bit of refactoring is done on MapActivity, now we store the exif data inside a list and it is passed as parameter inside the addMarkerCluster method.
The main problem to fix now is the multi drawable feature. Some lags appears and to prevent this is necessary to do some work in background but Ui doesn't follow.